ui: phase 5j — extract Confirm mode handler
Move handleConfirmMode (~14 lines: Esc/Enter forwarding to the
confirm prompt overlay, returning the prompt's registered Cmd on
result) to mode_confirm.go. Test shim added.

24/24 ui packages green.
